CAMEL-15655: Finish up the consumer
